Fix after self review.
